third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/embedded-content/the-img-element/update-the-image-data/fail-to-resolve-expected.txt

This is a testharness.js-based test.
FAIL <img src="//["> assert_equals: expected "//[" but got ""
FAIL <img srcset="//["> assert_equals: expected "//[" but got ""
FAIL <img srcset="//[" src="/images/red.png"> assert_equals: expected "//[" but got ""
FAIL <img srcset="//[, /images/red.png"> assert_equals: expected "//[" but got ""
Harness: the test ran to completion.
